Attempt to parse title from Markdown front matter
I'm slightly concerned about this, since I feel like it opens the door to trying
to parse titles out of all kinds of ill-defined formats, and I'd like this to be
the only non-Org one!

BUT: I sure do have a blog based on Jekyll and it'd be nice to see my post
titles in my search results.