Add interface check for docsim 0.1.3
`docsim` 0.1.4 will be updating the command-line interface in breaking ways, and
a few people might still have 0.1.3 installed, so it'd be nice to handle that
transition smoothly.

Realistically we ought to be able to remove this check in a week or two; I just
don't want to scare people off while the project's brand new. =)