Allow users to redefine note title parsing
We try to display the "title" of a file, not just its relative path, in the
results buffer. Currently, that means:

- If the file has a `.org` extension and a `#+TITLE:` keyword, use that value.
- Otherwise, if it's got YAML front matter at the top with a `title:` key, as
  with Jekyll blog posts in Markdown, use that value.
- If neither of those is available, just show the relative path to the file.

These probably won't satisfy everyone, so this provides a mechanism by which
users can define their own title parsing functions.

If the default titles aren't working for you, you can write your own and point
`docsim-get-title-function` to it. Your function should take an absolute file
path as an argument and return a string that represents the title of that file.

Closes #4.